FIX dependencies, scripts, readme

+++ run_demo.sh
@@ -14,7 +14,12 @@ echo "extraction done\n"
 python3 src/getTerrassesPizzas.py
 python3 src/prepare_token.py
 bash clean_tokens.sh
-rm src/NLP/data/toPredict/*.txt
+if [[ -d src/NLP/data/toPredict ]]
+then
+    rm src/NLP/data/toPredict/*.txt
+else
+    mkdir src/NLP/data/toPredict
+fi
 cp data/tokens/* src/NLP/data/toPredict/
 cd src/NLP/
 bash run.sh -f ../../data/tokens/ -t no -p yes
